Fixed PY2 test
